pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "project-starter"
version = "0.0.1"
description = "Starter repository for student's projects for info 2602"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Kwasi Edwards", email = "kwasiedwards@gmail.com" }
]
dependencies = [
    "asyncpg",
    "fastapi[all]",
    "google-adk",
    "httpx",
    "itsdangerous",
    "jinja2",
    "litellm",
    "psycopg2-binary",
    "pwdlib[argon2]",
    "pyjwt",
    "pytest",
    "pytest-asyncio",
    "python-jose",
    "ruff",
    "sqlmodel",
    "tabulate",
    "typer",
    "uvicorn[standard]",
]
[project.optional-dependencies]
dev = [

]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools]
packages = ["app"]
